src/requirements.txt
flask==2.3.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
ultralytics>=8.0.0
opencv-python>=4.7.0
pillow>=9.5.0
pydicom>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
